# app/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Focus flow app API",
        "description": "",
        "license": {
            "name": ""
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/category/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get a category by id",
                "operationId": "get_category",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetCategoryResponseDto"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Category not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Category"
                ],
                "summary": "Update a category",
                "operationId": "update_category_api",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCategoryDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Category updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateCategoryResponseDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Category not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Delete a category by id",
                "description": "Delete a category and all its tasks by id",
                "operationId": "delete_categories_api",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteCategoriesResponseDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
        "/api/focus-sessions/{id}": {
            "put": {
                "tags": [
                    "Session"
                ],
                "summary": "Update focus session",
                "operationId": "update_session_api",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateFocusSessionDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "Session updated successfully"
                    },
                    "400": {
                        "description": "Bad request - validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "409": {
                        "description": "Session already exists"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
        "/api/tasks/{id}": {
            "put": {
                "tags": [
                    "Task"
                ],
                "summary": "Update a task",
                "operationId": "update_task_api",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task ID to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTaskDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Task updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateTaskResponseDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Task not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CategoryDto": {
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "color",
                    "tasks"
                ],
                "properties": {
                    "color": {
                        "type": "string"
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "tasks": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TaskDto"
                        }
                    }
                }
            },
...
            "DeleteCategoriesResponseDto": {
                "type": "object",
                "required": [
                    "deleted_ids"
                ],
                "properties": {
                    "deleted_ids": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "GetCategoryResponseDto": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/CategoryDto"
                    }
                ]
            },
...
            "TaskDto": {
                "type": "object",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "categoryId": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "completedAt": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int64"
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "scheduledDate": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int64"
                    }
                }
            },
...
            "UpdateCategoryDto": {
                "type": "object",
                "properties": {
                    "color": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                }
            },
            "UpdateCategoryResponseDto": {
                "type": "object",
                "required": [
                    "updatedCategory"
                ],
                "properties": {
                    "updatedCategory": {
                        "$ref": "#/components/schemas/CategoryDto"
                    }
                }
            },
            "UpdateFocusSessionDto": {
                "type": "object",
                "properties": {
                    "categoryId": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "concentrationScore": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int32"
                    },
                    "endedAt": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int64"
                    },
                    "notes": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "startedAt": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int64"
                    },
                    "taskId": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                }
            },
...
            "UpdateTaskDto": {
                "type": "object",
                "properties": {
                    "categoryId": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "completedAt": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int64"
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "scheduledDate": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int64"
                    }
                }
            },
            "UpdateTaskResponseDto": {
                "type": "object",
                "required": [
                    "updatedTask"
                ],
                "properties": {
                    "updatedTask": {
                        "$ref": "#/components/schemas/TaskDto"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}