# backend/crates/adapters/src/http/dto/common/category_dto.rs
use crate::http::dto::common::task_dto::TaskDto;
use serde::{Deserialize, Serialize};
use utoipa::ToSchema;

#[derive(Debug, Serialize, Deserialize, ToSchema)]
pub struct CategoryDto {
    pub id: String,
    pub name: String,
    pub description: Option<String>,
    pub color: String,
    pub tasks: Vec<TaskDto>,
}